API + doc API + handlers + repo level to RK-2 (#29)
* TP-046 update: infro migrations

* TP-046 update: person handler

* TP-046 update: filler for images film and person

* TP-046 add: collection by tag handler

* TP-046 update: filler

* TP-046 update: filler

* TP-046 update: filler full

* TP-046 update: Date

* TP-046 update: Date

* TP-72a (not working) add: csrf token, in progress csrf middleware

* TP-046 update: pgx

* TP-046 update: pgx

* TP-046 update: pgx for collections persons and profiles

* TP-046 update: pgx for collections persons and profiles

* TP-046 add: film handler in process

* TP-046 add: film handler in process

* TP-046 add: review handler

* TP-046 add: trigger on prod, update: filler, handlers

* TP-046 update: images

* TP-046 update: images

* TP-046 update: docker

* TP-046 update: docker

* TP-046 update: docker

* TP-046 update: docker

* TP-046 update: handlers

* TP-046 update: handlers

* TP-046 update: data and api

* TP-046 update: data and api

* TP-046 update: data and api

* TP-72a add: csrf middleware, update: handler interface, handlers

* Merge end

* add user to context in auth middleware

* fix: add user to context in auth middleware

* TP-5f2 update: auth refactoring

* TP-036 update: repo level

* TP-036 update: repo level

* TP-036 update: repo level

* TP-046 update: user handlers

* TP-036 update: user settings and images with access

* TP-5f2 naming fix

* TP-036 update: user settings and images with access

* TP-036 update: filler

* TP-046 add: new api doc

* TP-5f2 add: auth database support

* TP-046 add: new api doc

* TP-5f2 fix old comments

* TP-5f2 add: validation user data

* Merge end + new tasks

* init after merge

* TP-046 add: more logging, user settings put

* TP-046 add: more logging, user settings put

* TP-046 update: images

* TP-046 update: configs

* TP-046 update: configs

* TP-865 refactoring auth, update session

* TP-046 update: docker main conteiner

* TP-865 lint

* TP-046 update: auth + user

* TP-046 update: conf

* TP-046 update: auth + user

* TP-046 update: TLS configs

* TP-046 update: docker main conteiner

* TP-046 add: Rate RateDrop NewReview handlers

* TP-046 add: Rate RateDrop NewReview handlers

* TP-046 add: Rate RateDrop NewReview handlers all in test

* TP-046 add: Rate RateDrop NewReview handlers all in test

* TP-046 update: doc handlers

* TP-046 update validation

* TP-046 update validation

* TP-046 update: doc handlers

* TP-3ca add logs, xss-protection

* TP-3ca restore go.sum and go.mod files

* TP-3ca add bluemonday package

* TP-3ca lint

* TP-04a add: XSS sec in reviews

* TP-2a7 update: error wrap

* update logging

* fix script

* TP-04a add: API end

* TP update: migrations

* TP update: migrations

* TP update: migrations

* TP update: migrations

+++ internal/collection/delivery/models/shortcollection.go
@@ -5,10 +5,10 @@ import (
 )
 
 type ShortFilmCollectionResponse struct {
-	Name       string  `json:"name,omitempty" example:"Популярное"`
-	Poster     string  `json:"poster,omitempty" example:"42"`
-	CountLikes float32 `json:"count_likes,omitempty" example:"1023"`
-	CountFilms float32 `json:"count_films,omitempty"  example:"10"`
+	Name       string `json:"name,omitempty" example:"Популярное"`
+	Poster     string `json:"poster,omitempty" example:"42"`
+	CountLikes int    `json:"count_likes,omitempty" example:"1023"`
+	CountFilms int    `json:"count_films,omitempty"  example:"10"`
 }
 
 func NewShortFilmCollectionResponse(collection *models.Collection) *ShortFilmCollectionResponse {
